contrib/diksha/requirements.txt
openpyxl==3.0.5
python-slugify==4.0.1
gdown==3.12.2
zimscraperlib==1.3.4
